수정시 password verify 오류 수정
details 글 수정 아이콘 클릭시 update 페이지로 이동
details 글 삭제 아이콘 클릭시 서버에 delete 요청 구현
아직 delete에서 password verify하는 기능은 구현 안됨
